Import Upstream version 1.6.38

+++ t/lib/MemcachedTest.pm
@@ -264,6 +264,7 @@ sub run_help {
 
 # -1 if the pid is actually dead.
 sub is_running {
+    return unless defined $_[0];
     return waitpid($_[0], WNOHANG) >= 0 ? 1 : 0;
 }
 
@@ -273,6 +274,10 @@ sub new_memcached {
     my $host = '127.0.0.1';
     my $ssl_enabled  = enabled_tls_testing();
     my $unix_socket_disabled  = !supports_unix_socket();
+    my $use_external = 0;
+    if ($ENV{T_MEMD_EXTERNAL}) {
+        $use_external = $ENV{T_MEMD_EXTERNAL};
+    }
 
     if ($ENV{T_MEMD_USE_DAEMON}) {
         my ($host, $port) = ($ENV{T_MEMD_USE_DAEMON} =~ m/^([^:]+):(\d+)$/);
@@ -322,29 +327,39 @@ sub new_memcached {
     } elsif ($args !~ /-s (\S+)/) {
         my $num = @unixsockets;
         my $file = "/tmp/memcachetest.$$.$num";
+        if ($use_external) {
+            $file = "/tmp/memcachedtest.$use_external.$num";
+        }
         $args .= " -s $file";
         push(@unixsockets, $file);
     }
 
-    my $childpid = fork();
-
+    my $wait_tries = 60;
     my $exe = get_memcached_exe();
+    my $childpid;
+    if ($use_external) {
+        print STDERR "External daemon requested. Start arguments:\n$exe $args\n";
+        $wait_tries = 10000;
+        $childpid;
+    } else {
+        $childpid = fork();
 
-    unless ($childpid) {
-        my $valgrind = "";
-        my $valgrind_args = "--quiet --error-exitcode=1 --exit-on-first-error=yes";
-        if ($ENV{VALGRIND_ARGS}) {
-            $valgrind_args = $ENV{VALGRIND_ARGS};
+        unless ($childpid) {
+            my $valgrind = "";
+            my $valgrind_args = "--quiet --error-exitcode=1 --exit-on-first-error=yes";
+            if ($ENV{VALGRIND_ARGS}) {
+                $valgrind_args = $ENV{VALGRIND_ARGS};
+            }
+            if ($ENV{VALGRIND_TEST}) {
+                $valgrind = "valgrind $valgrind_args";
+                # NOTE: caller file stuff.
+                $valgrind .= " $ENV{VALGRIND_EXTRA_ARGS}";
+            }
+            my $cmd = "$builddir/timedrun 600 $valgrind $exe $args";
+            #print STDERR "RUN: $cmd\n\n";
+            exec $cmd;
+            exit; # never gets here.
         }
-        if ($ENV{VALGRIND_TEST}) {
-            $valgrind = "valgrind $valgrind_args";
-            # NOTE: caller file stuff.
-            $valgrind .= " $ENV{VALGRIND_EXTRA_ARGS}";
-        }
-        my $cmd = "$builddir/timedrun 600 $valgrind $exe $args";
-        #print STDERR "RUN: $cmd\n\n";
-        exec $cmd;
-        exit; # never gets here.
     }
 
     # unix domain sockets
@@ -352,8 +367,7 @@ sub new_memcached {
         # A slow/emulated/valgrinded/etc system may take longer than a second
         # for the unix socket to appear.
         my $filename = $1;
-        for (1..20) {
-            sleep 1;
+        for (1..$wait_tries) {
             my $conn = IO::Socket::UNIX->new(Peer => $filename);
 
             if ($conn) {
@@ -363,8 +377,10 @@ sub new_memcached {
                                               host => $host,
                                               port => $port);
             } else {
-                croak("Failed to connect to unix socket: memcached not running") unless is_running($childpid);
-                sleep 1;
+                if (!$ENV{T_MEMD_EXTERNAL}) {
+                    croak("Failed to connect to unix socket: memcached not running") unless is_running($childpid);
+                }
+                select undef, undef, undef, 0.20;
             }
         }
         croak("Failed to connect to unix domain socket: $! '$filename'") if $@;
@@ -414,23 +430,48 @@ sub new {
 
 sub DESTROY {
     my $self = shift;
-    kill 2, $self->{pid};
+    if ($self->{pid}) {
+        kill 2, $self->{pid};
+    } else {
+        print STDERR "WANT TO ISSUE KILL: 2\n";
+    }
 }
 
 sub stop {
     my $self = shift;
-    kill 15, $self->{pid};
+    if ($self->{pid}) {
+        kill 15, $self->{pid};
+    } else {
+        print STDERR "WANT TO ISSUE KILL: 15\n";
+    }
 }
 
 sub graceful_stop {
     my $self = shift;
-    kill 'SIGUSR1', $self->{pid};
+    if ($self->{pid}) {
+        kill 'SIGUSR1', $self->{pid};
+    } else {
+        print STDERR "WANT TO ISSUE KILL: SIGUSR1\n";
+    }
+}
+
+sub reload {
+    my $self = shift;
+    if ($self->{pid}) {
+        kill 'SIGHUP', $self->{pid};
+    } else {
+        print STDERR "WANT TO ISSUE KILL: SIGHUP\n";
+    }
 }
 
 # -1 if the pid is actually dead.
 sub is_running {
     my $self = shift;
-    return waitpid($self->{pid}, WNOHANG) >= 0 ? 1 : 0;
+    if ($self->{pid}) {
+        return waitpid($self->{pid}, WNOHANG) >= 0 ? 1 : 0;
+    } else {
+        print STDERR "WANTED TO CHECK IF DAEMON IS RUNNING\n";
+    }
 }
 
 sub host { $_[0]{host} }
@@ -465,6 +506,19 @@ sub new_sock {
     }
 }
 
+# needed for a specific test
+sub new_nocert_tls_sock {
+    my $self = shift;
+    if (MemcachedTest::enabled_tls_testing()) {
+        my $port = shift;
+        my $ssl_version = shift;
+        return eval qq{ IO::Socket::SSL->new(PeerAddr => "$self->{host}:$port",
+                                    SSL_verify_mode => IO::Socket::SSL::SSL_VERIFY_NONE,
+                                    SSL_version => '$ssl_version');
+                                    };
+    }
+}
+
 sub new_udp_sock {
     my $self = shift;
     return IO::Socket::INET->new(PeerAddr => '127.0.0.1',
@@ -474,6 +528,226 @@ sub new_udp_sock {
                                  LocalPort => MemcachedTest::free_port('udp'),
         );
 
+}
+
+############################################################################
+package Memcached::ProxyTest;
+use IO::Socket qw(AF_INET SOCK_STREAM);
+# We call out to Test::More because of some package instancing. Not completely
+# sure if this is necessary anymore?
+use strict;
+use warnings;
+
+sub new {
+    my ($class, %p) = @_;
+
+    die "needs servers argument"
+        unless exists $p{servers} && ref($p{servers}) eq 'ARRAY';
+
+    $p{_srv} = [];
+    for my $port (@{$p{servers}}) {
+        my $srv = _mock_server($port);
+        Test::More::ok(defined $srv, "mock server object created");
+        push(@{$p{_srv}}, $srv);
+    }
+
+    $p{_csel} = IO::Select->new();
+
+    return bless \%p, $class;
+}
+
+sub _mock_server {
+    my $port = shift;
+    my $srv = IO::Socket->new(
+        Domain => AF_INET,
+        Type => SOCK_STREAM,
+        Proto => 'tcp',
+        LocalHost => '127.0.0.1',
+        LocalPort => $port,
+        ReusePort => 1,
+        Listen => 5) || die "IO::Socket: $@";
+    return $srv;
+}
+
+sub _accept_backend {
+    my $srv = shift;
+    my $be = $srv->accept();
+    $be->autoflush(1);
+    Test::More::ok(defined $be, "mock backend created");
+    Test::More::like(<$be>, qr/version/, "received version command");
+    my $sent = $be->send("VERSION 1.0.0-mock\r\n");
+    Test::More::cmp_ok($sent, '>', 0, "wrote VERSION back to backend");
+
+    return $be;
+}
+
+sub accept_backends {
+    my $self = shift;
+    $self->{_be} = [];
+    for my $srv (@{$self->{_srv}}) {
+        my $be = _accept_backend($srv);
+        push(@{$self->{_be}}, $be);
+    }
+}
+
+sub accept_backend {
+    my $self = shift;
+    my $idx = shift;
+    $self->{_be}->[$idx] = _accept_backend($self->{_srv}->[$idx]);
+}
+
+sub close_backend {
+    my $self = shift;
+    my $idx = shift;
+    $self->{_be}->[$idx]->close;
+}
+
+sub srv_accept_waiting {
+    my $self = shift;
+    my $idx = shift;
+    my $wait = shift || 1;
+    my $s = IO::Select->new();
+    $s->add($self->{_srv}->[$idx]);
+    return $s->can_read($wait);
+}
+
+sub set_c {
+    my $self = shift;
+    my $sel = $self->{_csel};
+    if (exists $self->{_c}) {
+        $sel->remove($self->{_c});
+    }
+    $self->{_c} = shift;
+    $sel->add($self->{_c});
+}
+
+sub check_c {
+    my $self = shift;
+    my $c = $self->{_c};
+    print $c "version\r\n";
+    Test::More::like(scalar <$c>, qr/VERSION /, "version received");
+}
+
+sub wait_c {
+    my ($self, $wait) = @_;
+    return $self->{_csel}->can_read($wait);
+}
+
+# Remembers the last command sent to the client socket.
+sub c_send {
+    my $self = shift;
+    my $cmd = shift;
+    my $c = $self->{_c};
+    print $c $cmd;
+    $self->{_cmd} = $cmd;
+}
+
+# Backends can be specified as a bare number, and array reference ([0,1,2]),
+# or 'all' to run against all available backends.
+sub _be_list {
+    my $self = shift;
+    my $list = shift;
+    my @l = ();
+    if (ref $list eq '') {
+        if (exists $self->{_be}->[$list]) {
+            push(@l, $self->{_be}->[$list]);
+        } elsif ($list eq 'all') {
+            @l = @{$self->{_be}};
+        } else {
+            die "unknown argument";
+        }
+    } elsif (ref $list eq 'ARRAY') {
+        for my $i (@$list) {
+            push(@l, $self->{_be}->[$i]);
+        }
+    }
+    return \@l;
+}
+
+# Check that the last command sent to the client arrives at the backends.
+# This is a common case so this saves typing/errors while writing tests.
+sub be_recv_c {
+    my $self = shift;
+    my $list = shift;
+    my $detail = shift || 'be received data';
+    die "issue a a command with c_send before calling be_recv_c" unless exists $self->{_cmd};
+
+    my $l = $self->_be_list($list);
+    my $cmd = $self->{_cmd};
+    my @cmds = split(/(?<=\r\n)/, $cmd);
+    for my $be (@$l) {
+        for my $c (@cmds) {
+            Test::More::is(scalar <$be>, $c, $detail);
+        }
+    }
+}
+
+sub be_recv_like {
+    my $self = shift;
+    my $list = shift;
+    my $cmd = shift || die "must provide a command to check";
+    my $detail = shift || 'be received data';
+
+    my $l = $self->_be_list($list);
+    for my $be (@$l) {
+        Test::More::like(scalar <$be>, $cmd, $detail);
+    }
+}
+
+# Receive a different/specific string to the backend socket.
+sub be_recv {
+    my $self = shift;
+    my $list = shift;
+    my $cmd = shift || die "must provide a command to check";
+    my $detail = shift || 'be received data';
+
+    my $l = $self->_be_list($list);
+    for my $be (@$l) {
+        Test::More::is(scalar <$be>, $cmd, $detail);
+    }
+}
+
+# Sends a specific command to a backend socket back towards the proxy.
+# Remembers the last command sent.
+sub be_send {
+    my $self = shift;
+    my $list = shift;
+    my $cmd = shift || die "must provide data to return";
+
+    $self->{_becmd} = $cmd;
+    my $l = $self->_be_list($list);
+    for my $be (@$l) {
+        print $be $cmd;
+    }
+}
+
+# Client receives the last command sent to any backend. This is also a common
+# case so we save some typing/errors by remembering this here.
+sub c_recv_be {
+    my $self = shift;
+    my $detail = shift || 'client received be response';
+
+    my $cmd = $self->{_becmd};
+    my $c = $self->{_c};
+    Test::More::is(scalar <$c>, $cmd, $detail);
+}
+
+# Client to receive an arbitrary string.
+sub c_recv {
+    my $self = shift;
+    my $cmd = shift;
+    my $detail = shift || 'client received response';
+
+    my $c = $self->{_c};
+    Test::More::is(scalar <$c>, $cmd, $detail);
+}
+
+# Clear out any remembered commands and check the client pipe is clear.
+sub clear {
+    my $self = shift;
+    delete $self->{_becmd} if exists $self->{_becmd};
+    delete $self->{_cmd} if exists $self->{_cmd};
+    $self->check_c();
 }
 
 1;
